+++ src/main/java/domain/menu/MenuRepository.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.menu;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
