+++ src/main/java/domain/table/TableRepository.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.table;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
